rename cachemanager
