+++ trinity/manager/__init__.py
@@ -1,4 +1,4 @@
-from trinity.manager.manager import StateManager
+from trinity.manager.state_manager import StateManager
 from trinity.manager.synchronizer import Synchronizer
 
 __all__ = [
