+++ trinity/manager/state_manager.py
